Added distance-based crosshair scaling option
This adds an option to change crosshair size based on distance.

Also, removed LAF flags being passed to LineTrace. I assume this is an oversight and at some point LineAttack was used instead of LineTrace. LineTrace does not utilize LAF flags, and this could lead to wrong flags being unintentionally passed to LineTrace.